requirements.txt
Django==3.0.6
djangorestframework==3.11.0
psycopg2-binary==2.8.5
drf-yasg==1.17.1